Move *MetaData classes to meta subpackage, rename MetaData to ExchangeMetaData
