+++ xchange-core/src/main/java/com/xeiam/xchange/ExchangeSpecification.java
@@ -3,8 +3,6 @@ package com.xeiam.xchange;
 import java.util.HashMap;
 import java.util.Map;
 
-import com.xeiam.xchange.dto.MetaData;
-
 /**
  * <p>
  * Specification to provide the following to {@link ExchangeFactory}:
@@ -369,7 +367,7 @@ public class ExchangeSpecification {
   }
 
   /**
-   * Get the override file for generating the {@link MetaData} object. By default, the {@link MetaData} object is loaded at startup from a json file
+   * Get the override file for generating the {@link com.xeiam.xchange.dto.meta.ExchangeMetaData} object. By default, the {@link com.xeiam.xchange.dto.meta.ExchangeMetaData} object is loaded at startup from a json file
    * on the classpath with the same name as the name of the exchange as defined in {@link ExchangeSpecification}. With this parameter, you can
    * override that file with a file of your choice located outside of the classpath.
    *
@@ -381,7 +379,7 @@ public class ExchangeSpecification {
   }
 
   /**
-   * Set the override file for generating the {@link MetaData} object. By default, the {@link MetaData} object is loaded at startup from a json file
+   * Set the override file for generating the {@link com.xeiam.xchange.dto.meta.ExchangeMetaData} object. By default, the {@link com.xeiam.xchange.dto.meta.ExchangeMetaData} object is loaded at startup from a json file
    * on the classpath with the same name as the name of the exchange as defined in {@link ExchangeSpecification}. With this parameter, you can
    * override that file with a file of your choice located outside of the classpath.
    *
