+++ xchange-core/src/main/java/com/xeiam/xchange/dto/meta/MarketMetaData.java
@@ -1,4 +1,4 @@
-package com.xeiam.xchange.dto;
+package com.xeiam.xchange.dto.meta;
 
 import java.math.BigDecimal;
